refactor(agents): separate catalog source preparation
Move the unchanged catalog-source preparer into its scoped-catalog owner
and update build, worker, and test imports. Preserve lazy worker loading,
static and live discovery, read-only planning, and snapshot capture order.

Related: #140674
